[MCUX-80794] Fix errors and warnings
qoi.h: Fix invalid digit in decimal constant error.
vglite_layer.h: Fix integer conversion resulted in
a change of sign warning.

+++ example/vglite_layer/vglite_layer.h
@@ -96,7 +96,7 @@ typedef struct stroke_info {
 
 typedef struct msg_glyph_item {
     uint16_t g_u16; /* Unicode number of glyph */
-    uint16_t k; /* kerning difference between g1 -> g2 transition */
+    int32_t k; /* kerning difference between g1 -> g2 transition */
 } msg_glyph_item_t;
 
 typedef struct svg_text_string_data {
